decoupled sklearn preprocessing from text preprocessing with heavy dependencies
